infra/functiongraph/chatbot/requirements.txt
chromadb==1.0.5
fastapi==0.115.9
langchain==0.3.23
langid==1.1.6
librosa>=0.10.1,<0.11
pandas==2.2.1
psycopg2_binary==2.9.10
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.9
requests==2.32.3
sentence_transformers==4.0.2
torch==2.7.0
transformers==4.51.2
uvicorn==0.29.0
